+++ wxWidgets-tree-list/tools/redhat/build_debug.sh
@@ -1,0 +1,2 @@
+rm build/redhat-debug/main
+cmake --build build/redhat-debug