+++ wxWidgets-tree-list/tools/redhat/build_release.sh
@@ -1,0 +1,1 @@
+cmake --build build/redhat-release