+++ wxWidgets-tree-list/tools/redhat/clean.sh
@@ -1,0 +1,2 @@
+cmake --build build/redhat-debug --target clean
+cmake --build build/redhat-release --target clean