+++ wxWidgets-tree-list/tools/redhat/configure_cmake_debug.sh
@@ -1,0 +1,3 @@
+rm -rf build/redhat-debug
+
+cmake -S . -B build/redhat-debug -DCMAKE_BUILD_TYPE=Debug -DREDHAT=1