+++ wxWidgets-tree-list/tools/redhat/configure_cmake_release.sh
@@ -1,0 +1,3 @@
+rm -rf build/redhat-release
+
+cmake -S . -B build/redhat-release -DCMAKE_BUILD_TYPE=Release -DREDHAT=1