+++ wxWidgets-tree-list/tools/redhat/run_release.sh
@@ -1,0 +1,1 @@
+build/redhat-release/main