+++ HISTORY.md
@@ -2,3 +2,4 @@
  - 2023-05-30 - Created wxWidgets Repository
  - 2023-05-30 - wxWidgets-hello-world: Base wxWidgets project configured for vcpkg and cmake
  - 2023-06-03 - wxWidgets-tree-list: Enhanced base wxWidgets project configured for vcpkg and cmake. Adds toolbars, status bar, splitters, tree and list controls.
+ - 2023-06-13 - wxWidgets-tree-list: Added support for Red Hat Enterprise 9.2.